pyproject.toml
[project]
name = "face-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "cmake>=3.25.0",
    "dlib-bin>=20.0.0",
    "fastapi>=0.131.0",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "huggingface-hub==0.14.1",
    "imgbeddings>=0.1.0",
    "ipython>=9.10.0",
    "opencv-python>=4.13.0.92",
    "psycopg2-binary>=2.9.11",
    "pymssql>=2.3.13",
    "pyodbc>=5.3.0",
    "pytest>=9.0.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.41.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
